pyproject.toml
[project]
name = "eth-wallet-tracker"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "pandas>=2.1.0",
    "web3>=6.12.0",
    "aiohttp>=3.9.0",
    "python-dateutil>=2.8.2",
    "click>=8.1.0"
]

[project.scripts]
eth-tracker = "eth_wallet_tracker.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eth_wallet_tracker"]
